+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [PR #2356](https://github.com/rqlite/rqlite/pull/2356): Consistent level-to-string conversions.
 - [PR #2358](https://github.com/rqlite/rqlite/pull/2358): Factor out Protobuf Consistency levels.
 - [PR #2357](https://github.com/rqlite/rqlite/pull/2357): Case insensitive BackupFormat query param.
+- [PR #2359](https://github.com/rqlite/rqlite/pull/2359): Don't modify supplied Query Level.
 
 ## v9.1.2 (October 3rd 2025)
 There are no changes in this release relative to v9.1.1. However this release is the first that mirrors container images to the [GitHub Container Registry](https://docs.github.com/en/packages).